requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
pillow==10.2.0
ipython==8.21.0
python-dotenv==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
coverage==7.4.4
requests==2.31.0
redis==5.0.3
celery==5.3.6
django-celery-beat==2.6.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
setuptools==69.5.1